rpc: handle EAGAIN when SSL_ERROR_SYSCALL is returned
Problem:
A return value of ENODATA was forcibly returned in the case where
SSL_get_error(r) returned SSL_ERROR_SYSCALL. Sometimes SSL_ERROR_SYSCALL
is a transient error which is identified by setting errno to EAGAIN.
EAGAIN is not a fatal error and indicates that the syscall needs to be
retried.

Solution:
Bubble up the errno in case SSL_get_error(r) returns SSL_ERROR_SYSCALL
and let the upper layers handle it appropriately.

mainline:
> Reviewed-on: https://review.gluster.org/c/glusterfs/+/20993
> fixes: bz#1622405
> Change-Id: I76eff278378930ee79abbf9fa267a7e77356eed6
> BUG: 1622405

Change-Id: I76eff278378930ee79abbf9fa267a7e77356eed6
BUG: 1622308
Signed-off-by: Milind Changire <mchangir@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/154868
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>