socket: set 42 as default tpc-user-timeout
> Upstream: https://review.gluster.org/21170
> BUG: 1628605
> Change-Id: Ib8ad7c4ac6aac725b01a78f8c3d10cf4063d2ee6

The 'tcp-user-timeout' option is define in the 'socket' module, but it's
configured in 'protocol/server' and 'protocol/client', which are the
parents of the 'socket' module.

However, current options management logic only takes into consideration
default values specified in the 'socket' module itself, ignoring values
defined in the owner xlator.

This patch simply sets the default value of tcp-user-timeout in the
'socket' module so that server and client use the expected value.

Change-Id: Ib8ad7c4ac6aac725b01a78f8c3d10cf4063d2ee6
BUG: 1623874
Signed-off-by: Xavi Hernandez <xhernandez@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/150699
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>