extras: Add new options to group "virt"
> Upstream: https://review.gluster.org/21222
> BUG: 1630798
> Change-Id: I231db309de0e37c79cd44f5666da4cd776fefa04

In some of the recent performance tests on gluster-as-vm-image-store
use-case, it has been observed that sometimes the lone fuse thread can
hit near-100% CPU utilization and become a performance bottleneck.
Enabling client-io-threads (in addition to bumping up epoll threads on
server and client side) has shown to be helpful in getting around this
bottleneck and pushing more IOPs.

Change-Id: I231db309de0e37c79cd44f5666da4cd776fefa04
BUG: 1619627
Signed-off-by: Krutika Dhananjay <kdhananj@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/150696
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>