glusterd: don't wait for blockers flag for stop volume
With commit 4f6ae8 even though the overall transaction time for gluster
volume stop can be reduced, but based on testing it can't be guaranteed
that the transaction can finish in 3 minutes before the unlock timer
gets kicked in. The ground problem to this is the command serialization
which atomic field 'blockers' does for volume stop operation.

This patch removes that dependency for volume stop not to wait for
blockers.

> Change-Id: Ifaf120115dc7ed21cf21e65e25c4ec7c61ab8258
> Fixes: bz#1631128
> Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
> (Cherry picked from commit f72f18b3bf7f9535827a059a5090476b3141723f)
> (Reviewed on upstream link https://review.gluster.org/#/c/glusterfs/+/21242/)

Change-Id: Ia3cb4d812ea1c633b7a501a03e0bf25a20b45a8e
BUG: 1631329
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/151065
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>