glusterd: ignore RPC events when glusterd is shutting down
When glusterd receives a SIGTERM while it receives RPC
connect/disconnect/destroy events, the thread might lead to a crash
while accessing rcu_read_lock () as the clean up thread might have
already freed up the resources. This is more observable when glusterd
comes up with upgrade mode = on during upgrade process.

The solution is to ignore these events if glusterd is already in the
middle of cleanup_and_exit ().

> upstream patch : https://review.gluster.org/#/c/glusterfs/+/21330/

>Fixes: bz#1635593
>Change-Id: I12831d31c2f689d4deb038b83b9421bd5cce26d9
>Signed-off-by: Atin Mukherjee <amukherj@redhat.com>

Change-Id: I12831d31c2f689d4deb038b83b9421bd5cce26d9
BUG: 1635071
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/154848
Reviewed-by: Sanju Rakonde <srakonde@redhat.com>
Tested-by: Sanju Rakonde <srakonde@redhat.com>
Tested-by: RHGS Build Bot <nigelb@redhat.com>