mgmt/glusterd: use proper path to the volfile
    > Upstream: https://review.gluster.org/#/c/glusterfs/+/21314/
    > BUG: 1635050
    > Change-Id: I28b2dfa5d9b379fe943db92c2fdfea879a6a594e

NOTE: This patch is actually directly applied from the patch that
      was sent to the release-4.1 branch. The master branch patch
      will have merge conflicts due to the clang format changes done
      there. This is the patch which this commit is a backport of.

      upstream(4.1): https://review.gluster.org/#/c/glusterfs/+/21348/
      Chane-ID: I28b2dfa5d9b379fe943db92c2fdfea879a6a594e

Till now, glusterd was generating the volfile path for the snapshot
volume's bricks like this.

/snaps/<snap name>/<brick volfile>

But in reality, the path to the brick volfile for a snapshot volume is

/snaps/<snap name>/<snap volume name>/<brick volfile>

The above workaround was used to distinguish between a mount command used
to mount the snapshot volume, and a brick of the snapshot volume, so that
based on what is actually happening, glusterd can return the proper volfile
(client volfile for the former and the brick volfile for the latter). But,
this was causing problems for snapshot restore when brick multiplexing is
enabled. Because, with brick multiplexing, it tries to find the volfile
and sends GETSPEC rpc call to glusterd using the 2nd style of path i.e.

/snaps/<snap name>/<snap volume name>/<brick volfile>

So, when the snapshot brick (which is multiplexed) sends a GETSPEC rpc
request to glusterd for obtaining the brick volume file, glusterd was
returning the client volume file of the snapshot volume instead of the
brick volume file.

Change-Id: I28b2dfa5d9b379fe943db92c2fdfea879a6a594e
BUG: 1636291
Signed-off-by: Raghavendra Bhat <raghavendra@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/155129
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>