locks: handle "clear locks" xattr in fgetxattr too
The lock clearing procedure was kicked in only in
getxattr context. We need it to work the same way
if it's triggered via fgetxattr (as is the case
with interrupt handling).

Also cleaned up the instrumentation a bit (more logs,
proper management of allocated data).

Upstream: https://review.gluster.org/21820
> updates: #465
> Change-Id: Icfca26ee181da3b8e15ca3fcf61cd5702e2730c8
> Signed-off-by: Csaba Henk <csaba@redhat.com>

Change-Id: Ia15108fd6d92ea2bdb73cea5fb04126785b19663
BUG: 1595246
Signed-off-by: Csaba Henk <csaba@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/162551
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Amar Tumballi Suryanarayan <amarts@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>