ganesha-ha: ensure pacemaker is enabled after setup
There appears to be a race between `pcs cluster setup ...` early
in the setup and the `systemctl enable pacemaker` at the end. The
`pcs cluster setup ...` disables pacemaker and corosync. (Now, in
pacemaker-1.1.18. Was it always the case?)

I am not able to reproduce this on my devel system. I speculate that
on a busy system that the `pcs cluster setup ...` disable may, under
the right conditions, not run until after the setup script enables
it. It must require the right alignment of the Sun, Moon, and all
the planets.

Regardless, we'll use the --enable option to `pcs cluster setup ...`
to ensure that the cluster (re)starts pacemaker.

Label: DOWNSTREAM ONLY

Change-Id: I771ff62c37426438b80e61651a8b4ecaf2d549c3
BUG: 1637564
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/158294
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Soumya Koduri <skoduri@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>