cluster/dht: sync brick root perms on add brick
If a single brick is added to the volume and the
newly added brick is the first to respond to a
dht_revalidate call, its stbuf will not be merged
into local->stbuf as the brick does not yet have
a layout. The is_permission_different check therefore
fails to detect that an attr heal is required as it
only considers the stbuf values from existing bricks.
To fix this, merge all stbuf values into local->stbuf
and use local->prebuf to store the correct directory
attributes.

upstream patch: https://review.gluster.org/#/c/glusterfs/+/21611/

> Change-Id: Ic9e8b04a1ab9ed1248b6b056e3450bbafe32e1bc
> fixes: bz#1648298
> Signed-off-by: N Balachandran <nbalacha@redhat.com>

Change-Id: I329ce48555d15f741d7247290e749f0800d12df8
BUG: 1648296
Signed-off-by: N Balachandran <nbalacha@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/158493
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Susant Palai <spalai@redhat.com>
Reviewed-by: Raghavendra Gowdappa <rgowdapp@redhat.com>