cluster/afr: Do not update read_subvol in inode_ctx after rename/link fop
Since rename/link fops on a file will not change any data in it, it should
not update the read_subvol values in the inode_ctx, which interprets the
data & metadata readable subvols for that file. The old read_subvol values
should be retained even after the rename/link operations.

Backport of: https://review.gluster.org/#/c/glusterfs/+/21841

Change-Id: I22787938b3f8295a48162ab4f498e4010d66a1ab
BUG: 1645480
Signed-off-by: karthik-us <ksubrahm@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/158803
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Tested-by: Ravishankar Narayanankutty <ravishankar@redhat.com>
Reviewed-by: Ravishankar Narayanankutty <ravishankar@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>