mem-pool: add tracking of mem_pool that requested the allocation
This renames the current 'struct mem_pool' to 'struct mem_pool_shared'.
The mem_pool_shared is globally allocated and not specific for
particular objects.

A new 'struct mem_pool' gets allocated when mem_pool_new() is called. It
points to the mem_pool_shared that handles the actual allocation
requests. The 'struct mem_pool' is only used for accounting of the
objects that the caller requested and free'd.

All of these changes will be used to collect all the memory pools a
glusterfs_ctx_t is consuming, so that statedumps can be collected per
context.

> Updates: #307
> Change-Id: I6355d3f0251c928e0bbfc71be3431307c6f3a3da
> Signed-off-by: Niels de Vos <ndevos@redhat.com>
> Reviewed-on: https://review.gluster.org/18073
> Smoke: Gluster Build System <jenkins@build.gluster.org>
> CentOS-regression: Gluster Build System <jenkins@build.gluster.org>
> Reviewed-by: Amar Tumballi <amarts@redhat.com>
> Reviewed-by: Jeff Darcy <jeff@pl.atyp.us>
> Cherry picked from commit 2645e730b79b44fc035170657e43bb52f3e855c5

Change-Id: I6cce6284e4553c6ca59a90ad124c23c950db3148
BUG: 1648893
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>

Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>
Change-Id: I363d71152b1dd17eca53d9c327fcdf2f26c0fb61
Reviewed-on: https://code.engineering.redhat.com/gerrit/158930
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>