cluster/dht: Use percentages for space check
With heterogenous bricks now being supported in DHT
we could run into issues where files are not migrated
even though there is sufficient space in newly added bricks
which just happen to be considerably smaller than older
bricks. Using percentages instead of absolute available
space for space checks can mitigate that to some extent.

upstream patch:https://review.gluster.org/#/c/glusterfs/+/19101/
This is not an identical backport as there were some changes
to upstream master that are not available in the downstream code.

Marking bug-1247563.t bad as that used to depend on the easier
code to prevent a file from migrating. This will be removed
once we find a way to force a file migration failure.

Change-Id: Ie89bfdd114406a986b3ff4f53b0bb0fae6574c8e
BUG: 1290124
Signed-off-by: N Balachandran <nbalacha@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/159569
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Susant Palai <spalai@redhat.com>
Reviewed-by: Raghavendra Gowdappa <rgowdapp@redhat.com>